+++ cli/memory_custodian/__init__.py
@@ -1,4 +1,4 @@
 """MemoryCustodian CLI package."""
 
-__version__ = "0.6.0"
-__protocol_version__ = "0.4"
+__version__ = "0.7.0"
+__protocol_version__ = "0.5"